software/requirements.txt
cocotb==1.7.2
find-libpython==0.3.0
pip==23.1.2
pyserial==3.5
setuptools==59.6.0
wheel==0.37.1
